customer-device-api/requirements.txt
#requirements
flask==1.0.2
Flask-Cors==3.0.6
requests==2.19.1
more_itertools==4.2.0
paho-mqtt==1.4.0
smbus2==0.2.1
pyserial==3.4
pytz==2018.7
